UnitTest/Application/DirectorOperations/Queries/GetDirectorById/GetDirectorByIdQueryTests.cs
using System;
using System.Linq;
using AutoMapper;
using MovieStore.Application.DirectorOperations.Queries.GetDirectorById;
using MovieStore.DBOperations;
using MovieStore.Entities;
using FluentAssertions;
using TestSetup;
using Xunit;

namespace Application.DirectorOperations.Queries.GetDirectorById
{
  public class GetDirectorByIdQueryTests : IClassFixture<CommonTestFixture>
  {
    private readonly MovieStoreDbContext _context;
    private readonly IMapper _mapper;
    public GetDirectorByIdQueryTests(CommonTestFixture testFixture)
    {
      _context = testFixture.Context;
      _mapper = testFixture.Mapper;
    }

    [Fact]
    public void WhenGivenDirectorIsNotFound_Handle_ThrowsInvalidOperationException()
    {
      // arrange
      GetDirectorByIdQuery query = new GetDirectorByIdQuery(_context, _mapper);
      query.Id = 999;

      // act & assert
      FluentActions
        .Invoking(() => query.Handle())
        .Should().Throw<InvalidOperationException>()
        .And
        .Message.Should().Be("Yönetmen bulunamadı.");
    }

    [Fact]
    public void WhenValidInputsAreGiven_Director_ShouldBeReturned()
    {
      // arrange
      GetDirectorByIdQuery query = new GetDirectorByIdQuery(_context, _mapper);
      query.Id = 2;

      // act
      FluentActions.Invoking(() => query.Handle()).Invoke();

      // assert
      Director author = _context.Directors.SingleOrDefault(author => author.Id == query.Id);

      author.Should().NotBeNull();
    }
  }
}
